// src/app/pages/news/category/category.interface.ts
export interface ICategory {
  Children: [];
  AntiInjectionDate: string;
  AntiInjectionExpiredMinute: number;
  AntiInjectionRun: boolean;
  AntiInjectionToken: string;
  AntiInjectionTokenActionState: number;
  CreatedBy: number;
  CreatedDate: string;
  Description: string;
  Id: number;
  LinkParentIdNode: string;
  LinkSiteId: number;
  RecordStatus: number;
  Title: string;
  LinkParentId: number;
}